requirements.txt
lightrag-hku
tqdm
textract

accelerate
aioboto3
aiofiles
aiohttp
asyncpg

graspologic
gremlinpython
hnswlib
nano-vectordb
neo4j
networkx

numpy
ollama
openai
trulens
trulens-providers-openai
trulens-utils
oracledb
psycopg-pool
psycopg[binary,pool]
pydantic
pymilvus
pymongo
pymysql
python-dotenv
pyvis
setuptools
sqlalchemy
tenacity

tiktoken
torch
tqdm
transformers
xxhash